moved terminal-toolbar sub-component into Terminal.js
